Moved Analysis class to data structures module.
