+++ src/data_structures/analysis.py
